1. 修复职住分析部分路径硬编码问题 2. 文件都以parquet格式存储和读取

+++ sky-root/signal-pre/src/main/java/com/sky/signal/pre/processor/transportationhub/StationPersonClassification.java
@@ -99,7 +99,7 @@ public class StationPersonClassification implements Serializable {
         List<Row> resultList;
         Long sumMoveTime = 0L;
         for (Row row : rows) {
-            Integer moveTime = row.getAs("move_time");
+            Integer moveTime = (Integer) row.getAs("move_time");
             sumMoveTime += moveTime;
         }
         if (sumMoveTime >= ONE_HOUR) {
